Start to extract a `messages` crate.
The `common` crate extraction had some mixed results. It seems like it
can't compile some of the things included in that crate into wasm
making it the wrong thing to share between the frontend and backend. I
created a messages crate instead. My intention is for this to be the
type of all messages sent to and from the lambdas.